Remove the experiment with iterating notes and clean up the script in general
Upon further deliberation, it doesn't seem to make much sense to store the issues separately. It seems preferrable to store them all with the main branch tip commit on each run and replace that payload if it already exists (i.e. the trigger is not push but issues or workflow_dispatch).